test(graph): add E28.2 PR4 conformance suite + fix 14 pre-existing executor bugs
Closes the E28.2 chain. Validates that PgGraphExecutor and SnapshotGraphExecutor
return equivalent GraphResults for the same MoldPlan + workspace + revision pin,
covering path ordering, unordered neighbor sets, max_result_rows truncation,
max_path_count truncation, subgraph node enumeration, and unknown revision
errors. 10 conformance tests verde (7 PG-required, 3 unit-only).

Pre-existing bugs surfaced by the conformance harness and fixed:

  - pg_graph_executor: DISTINCT ON (path[last]) collapsed parallel paths per
    endpoint. Replaced with DISTINCT path + ORDER BY depth ASC, path ASC.
  - pg_graph_executor: execute_boolean pin propagation dropped caller revision
    (called execute(operand, (ws, RevisionId::NONE)) instead of threading
    caller pin through).
  - pg_graph_executor: execute_boolean Not semantics returned operand instead
    of universe - operand. Now queries graph_nodes directly via block_in_place.
  - pg_graph_executor: max_path_count truncation was missing (only
    max_result_rows was honored).
  - pg_graph_executor: neighbor query needed ORDER BY id ASC for stable order.
  - snapshot_graph_executor: edge labels used Debug form (e.g. "Calls")
    instead of Display "dependency.calls".
  - snapshot_graph_executor: nodes not sorted before LIMIT truncation, so PG
    and snapshot disagreed on which nodes get truncated.
  - snapshot_graph_executor: EdgeResult.properties was empty; now populated
    from CallGraph::all_dependencies_with_metadata().
  - m0010: notify_graph_change trigger referenced NEW.source_path on
    graph_edges inserts, but column was dropped after m0018. Fixed via
    TG_TABLE_NAME to conditionally include source_path only on graph_nodes
    inserts.
  - m0018/m0019 migration order swapped: m0019 unique index must exist before
    m0018 composite FKs reference it.
  - schema: "column" is reserved in PostgreSQL; SELECT column must be quoted.
  - call_graph: new all_dependencies_with_metadata() method returning
    (source, target, dep_type, provenance, confidence) for snapshot parity.
  - postgres_repository: confidence column is REAL (f32); cast to f64 to match
    ExtractionContext.
  - integration tests: #![cfg(feature = "postgres")] required for pg_test!
    macro expansion; pg_test! uses flavor = "multi_thread" so block_in_place
    in PgGraphExecutor::execute_with_limits is allowed.

Final test status (all verde):
  - 10/10 conformance
  - 12/12 pg_graph_executor
  - 16/16 snapshot_graph_executor
  - 1650/1650 cognicode-core lib + 27 ignored
  - 4/4 cognicode-explorer

+++ crates/cognicode-core/src/infrastructure/persistence/postgres_repository.rs
@@ -1356,7 +1356,11 @@ impl PostgresRepository {
             target_id: String,
             kind: String,
             provenance: String,
-            confidence: f64,
+            // Schema stores `confidence REAL` (f32). Cast to f64 to match
+            // the `GraphEdge.confidence: f64` field. Pre-existing schema
+            // mismatch surfaced when restarting the postgres container
+            // wiped the volume (2026-07-29).
+            confidence: f32,
         }
 
         let edges: Vec<GraphEdgeRow> = sqlx::query_as(
@@ -1395,7 +1399,8 @@ impl PostgresRepository {
 
             // Round-trip: provenance + confidence → ExtractionContext →
             // ConfidenceRules::assign → same (provenance, confidence).
-            let ctx = Self::provenance_to_extraction_context(provenance, row.confidence);
+            // Cast f32 → f64 to match `ExtractionContext`'s f64 confidence.
+            let ctx = Self::provenance_to_extraction_context(provenance, row.confidence as f64);
             graph
                 .add_dependency_with_provenance(src_id, tgt_id, dep_type, ctx)
                 .map_err(|e| {
